+++ internal/framework/services/userobjectstorage/data_sources.go
@@ -1,0 +1,10 @@
+package userobjectstorage
+
+import "github.com/hashicorp/terraform-plugin-framework/datasource"
+
+// DataSources returns the list of data sources for the userobjectstorage package.
+func DataSources() []func() datasource.DataSource {
+	return []func() datasource.DataSource{
+		NewBucketDataSource,
+	}
+}